Changing "Congressman" terminology to gender-neutral "Member of Congress"
The term "congressman" is a bit antiquated, and refers to a time when
all people holding government office were men. Even though there are
increasingly more women holding government office, language like this
can subtly reinforce the male dominance that continues to exist in
that field (and other fields too).